Port ArrayIntCopyTest
